Add DiffeomorphicVectorSpace and DiffeomorphisMatrixVectorSpace and use it to define NullRowSymmetricMatrices
